diameter: relay way back
During resend msg from irelay to orelay, information about receiver
compiles to msg. In this case we shoudn`t save local table to find
way back of the message.

Add function irelay:init/1 to add information to irelay about orelay
After that the relay can be tested in the cloud.